# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Planet Motors API
  description: |
    Planet Motors - Canada's Premier Online Used Car Marketplace
    
    ## Features
    - 210-point vehicle inspections
    - Multi-lender financing (6 Canadian banks)
    - Instant trade-in valuations via Canadian Black Book
    - Home delivery across Ontario
    - 10-day return policy
    
    ## Authentication
    Most endpoints require a Bearer token. Include it in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## Rate Limiting
    - Standard endpoints: 100 requests per 15 minutes
    - Financing endpoints: 10 requests per 15 minutes
  version: 1.0.0
  contact:
    name: Planet Motors API Support
    email: api@planetmotors.ca
    url: https://www.planetmotors.ca
  license:
    name: Proprietary
    url: https://www.planetmotors.ca/terms
servers:
  - url: https://api.planetmotors.ca/v1
    description: Production
  - url: https://api-staging.planetmotors.ca/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Local Development
...
paths:
...
  /orders:
    post:
      tags: [Orders]
      summary: Create order
      description: Create a new vehicle purchase order
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
    
    get:
      tags: [Orders]
      summary: List orders
      description: List customer orders
      operationId: listOrders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [created, pending, confirmed, preparing, delivering, delivered, returned]
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateOrderRequest:
      type: object
      required:
        - customerId
        - vehicleId
        - paymentMethod
      properties:
        customerId:
          type: string
        vehicleId:
          type: string
        financingOfferId:
          type: string
        paymentMethod:
          type: string
          enum: [financing, cash, bank_draft]
        tradeInOfferId:
          type: string
        deliveryType:
          type: string
          enum: [delivery, pickup]
        downPayment:
          type: number
    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            order:
              type: object
            message:
              type: string
            nextSteps:
              type: array
              items:
                type: string
    OrderListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            orders:
              type: array
              items:
                type: object
            pagination:
              $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean
...